+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- Add `favorite` operation. ([#86](https://github.com/turadg/pocket-ruby/pull/86))
+
 ## [0.4.0] - 2021-04-18
 
 - Fix Faraday error handling. ([#81](https://github.com/turadg/pocket-ruby/pull/81))